+++ src/webview/controller/javascript/source_config.ts
@@ -41,16 +41,9 @@ function main() {
 
   window.addEventListener('message', receive_message);
 
-  showAlert('Configuration reloaded.');
+  showAlert('Configuration reloaded.', 'success');
 }
 
-
-function showAlert(text: string) {
-  const box = document.getElementById('alertBox');
-  box.textContent = text;
-  box.style.display = 'block';
-  setTimeout(() => box.style.display = 'none', 2000);
-}