+++ src/webview/deployment/javascript/deployment_config.ts
@@ -4,6 +4,8 @@ import {
   Button
 } from "@vscode/webview-ui-toolkit";
 
+import { showAlert } from "../../tools/javascript/alertBox";
+
 // In order to use all the Webview UI Toolkit web components they
 // must be registered with the browser (i.e. webview) using the
 // syntax below.
@@ -24,15 +26,7 @@ function main() {
   
   window.addEventListener('message', receive_message);
 
-  showAlert('Configuration reloaded.');
-}
-
-
-function showAlert(text: string) {
-  const box = document.getElementById('alertBox');
-  box.textContent = text;
-  box.style.display = 'block';
-  setTimeout(() => box.style.display = 'none', 2000);
+  showAlert('Configuration reloaded.', 'success');
 }