+++ src/webview/source_list/javascript/source_list_config.ts
@@ -5,6 +5,7 @@ import {
 } from "@vscode/webview-ui-toolkit";
 
 import * as source from '../../../obi/Source';
+import { showAlert } from "../../tools/javascript/alertBox";
 
 // In order to use all the Webview UI Toolkit web components they
 // must be registered with the browser (i.e. webview) using the
@@ -42,15 +43,7 @@ function main() {
 
   window.addEventListener('message', receive_message);
 
-  showAlert('Configuration reloaded.');
-}
-
-
-function showAlert(text: string) {
-  const box = document.getElementById('alertBox');
-  box.textContent = text;
-  box.style.display = 'block';
-  setTimeout(() => box.style.display = 'none', 2000);
+  showAlert('Configuration reloaded.', 'success');
 }